add coparison table, change breakpoint to Int

+++ app/src/main/java/com/example/rentorbuy/data/baseUiState.kt
@@ -3,6 +3,6 @@ package com.example.rentorbuy.data
 data class baseUiState(
     val rentPrice: Float = 0f,
     val buyPrice: Float = 0.0f,
-    val breakPoint: Float = 0.0f,
+    val breakPoint: Int = 0,
     val modZero: Boolean = false
 )
